backend/requirements.txt
guardrails-ai>=0.5.0

# Hub validators — install via:
#   guardrails hub install hub://guardrails/detect_pii
#   guardrails hub install hub://guardrails/toxic_language
#   guardrails hub install hub://guardrails/detect_jailbreak
#   guardrails hub install hub://guardrails/secrets_present
#   guardrails hub install hub://guardrails/unusual_prompt
#   guardrails hub install hub://groundedai/grounded_ai_hallucination
#   guardrails hub install hub://guardrails/bias_check
